+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.17.2 May 14, 2025
+– Added GET check as fallback for PDF check as HEAD can be unreliable. Some servers (like IIS, older .NET apps) serve PDF content at non-.pdf URLs — like .ashx, .php, or extensionless URLs — while returning a valid PDF in the response. This attempts to resolve this issue.
+– Fixes PDF tag exclusion by adding typical brackets to the matching for PDF, MB and KB.
+
 ## 1.17.1 May 13, 2025
 – Augmented PDF link feature to insert PDF specific SVG into links including externally processed.
 – Added 'pdf' in title or class name checks on links to pass to proxy to allow redirect checking.